requirements.txt
streamlit
pandas
numpy
requests
pyyaml
pytest
